HeadToをInt値化

+++ app/src/main/java/com/example/stray_compass/MainActivity.kt
@@ -211,7 +211,7 @@ fun Viewer(
 
 @Composable
 fun Viewer(
-    azimuth: Int,
+    azimuth: Double,
     latitude: Double,
     longitude: Double,
     destination: Location,
